+++ fpdf/__init__.py
@@ -9,7 +9,7 @@ Gives direct access to some classes defined in submodules:
 * `fpdf.enums.YPos`
 * `fpdf.errors.FPDFException`
 * `fpdf.fonts.FontFace`
-* `fpdf.fpdf.TextStyle`
+* `fpdf.fonts.TextStyle`
 * `fpdf.prefs.ViewerPreferences`
 * `fpdf.template.Template`
 * `fpdf.template.FlexTemplate`
@@ -21,12 +21,11 @@ from .enums import Align, TextMode, XPos, YPos
 from .errors import FPDFException
 from .fpdf import (
     FPDF,
-    TextStyle,
     TitleStyle,
     FPDF_FONT_DIR as _FPDF_FONT_DIR,
     FPDF_VERSION as _FPDF_VERSION,
 )
-from .fonts import FontFace
+from .fonts import FontFace, TextStyle
 from .html import HTMLMixin, HTML2FPDF
 from .prefs import ViewerPreferences
 from .template import Template, FlexTemplate