+++ test/outline/test_outline.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from fpdf import FPDF, TextStyle, errors
+from fpdf import FPDF, TextStyle, TitleStyle, errors
 
 from test.conftest import assert_pdf_equal
 
@@ -107,6 +107,42 @@ def test_2_pages_outline(tmp_path):
     assert_pdf_equal(pdf, HERE / "2_pages_outline.pdf", tmp_path)
 
 
+def test_2_pages_outline_with_deprecated_TitleStyle(tmp_path):
+    pdf = FPDF()
+    pdf.set_font("Helvetica")
+    with pytest.warns(DeprecationWarning):
+        pdf.set_section_title_styles(
+            # Level 0 titles:
+            TitleStyle(
+                font_family="Times",
+                font_style="B",
+                font_size_pt=24,
+                color=128,
+                underline=True,
+                t_margin=10,
+                l_margin=10,
+                b_margin=0,
+            ),
+        )
+
+    pdf.add_page()
+    pdf.set_y(50)
+    pdf.set_font(size=40)
+    p(pdf, "Doc Title", align="C")
+    pdf.set_font(size=12)
+    pdf.insert_toc_placeholder(render_toc, pages=2)
+    for i in range(40):
+        pdf.start_section(f"Title {i}")
+        p(
+            pdf,
+            (
+                "Lorem ipsum dolor sit amet, consectetur adipiscing elit,"
+                " sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
+            ),
+        )
+    assert_pdf_equal(pdf, HERE / "2_pages_outline.pdf", tmp_path)
+
+
 def test_toc_with_nb_and_footer(tmp_path):  # issue-548
     class TestPDF(FPDF):
         def render_toc(self, outline):
@@ -189,6 +225,41 @@ def test_toc_with_font_style_override_bold(tmp_path):  # issue-1072
     assert_pdf_equal(pdf, HERE / "toc_with_font_style_override_bold2.pdf", tmp_path)
 
 
+def test_toc_without_font_style_with_deprecated_TitleStyle(tmp_path):
+    pdf = FPDF()
+    pdf.set_font("helvetica")
+    with pytest.warns(DeprecationWarning):
+        pdf.set_section_title_styles(
+            level0=TitleStyle(font_size_pt=28, l_margin=10), level1=TitleStyle()
+        )
+    pdf.add_page()
+    pdf.start_section("Title")
+    pdf.start_section("Subtitle", level=1)
+    assert_pdf_equal(pdf, HERE / "toc_without_font_style.pdf", tmp_path)
+
+
+def test_toc_with_font_style_override_bold_with_deprecated_TitleStyle(tmp_path):
+    pdf = FPDF()
+    pdf.add_page()
+    pdf.set_font("Helvetica", "B")
+    with pytest.warns(DeprecationWarning):
+        pdf.set_section_title_styles(
+            TitleStyle("Helvetica", font_size_pt=20, color=(0, 0, 0))
+        )
+    pdf.start_section("foo")
+    assert_pdf_equal(pdf, HERE / "toc_with_font_style_override_bold1.pdf", tmp_path)
+
+    pdf = FPDF()
+    pdf.add_page()
+    pdf.set_font("Helvetica", "B")
+    with pytest.warns(DeprecationWarning):
+        pdf.set_section_title_styles(
+            TitleStyle("Helvetica", font_style="", font_size_pt=20, color=(0, 0, 0))
+        )
+    pdf.start_section("foo")
+    assert_pdf_equal(pdf, HERE / "toc_with_font_style_override_bold2.pdf", tmp_path)
+
+
 def test_toc_with_table(tmp_path):  # issue-1079
     def render_toc_with_table(pdf: FPDF, outline: list):
         pdf.set_font(size=20)
@@ -264,6 +335,63 @@ def insert_test_content(pdf):
             b_margin=5,
         ),
     )
+
+    pdf.start_section("Title 1")
+    pdf.start_section("Subtitle 1.1", level=1)
+    p(
+        pdf,
+        (
+            "Lorem ipsum dolor sit amet, consectetur adipiscing elit,"
+            " sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
+        ),
+    )
+    pdf.add_page()
+    pdf.start_section("Subtitle 1.2", level=1)
+    p(
+        pdf,
+        "Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
+    )
+    pdf.add_page()
+    pdf.start_section("Title 2")
+    pdf.start_section("Subtitle 2.1", level=1)
+    p(
+        pdf,
+        "Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.",
+    )
+    pdf.add_page()
+    pdf.start_section("Subtitle 2.2", level=1)
+    p(
+        pdf,
+        "Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.",
+    )
+
+
+def insert_test_content_with_deprecated_TitleStyle(pdf):
+    with pytest.warns(DeprecationWarning):
+        pdf.set_section_title_styles(
+            # Level 0 titles:
+            TitleStyle(
+                font_family="Times",
+                font_style="B",
+                font_size_pt=24,
+                color=128,
+                underline=True,
+                t_margin=10,
+                l_margin=10,
+                b_margin=0,
+            ),
+            # Level 1 subtitles:
+            TitleStyle(
+                font_family="Times",
+                font_style="B",
+                font_size_pt=20,
+                color=128,
+                underline=True,
+                t_margin=10,
+                l_margin=20,
+                b_margin=5,
+            ),
+        )
 
     pdf.start_section("Title 1")
     pdf.start_section("Subtitle 1.1", level=1)